Rename wait_full_jitter to wait_random_exponential
The name "full jitter" can only really be interpreted in the context of
a particular blog post[1] that compares it to other possible algorithms.
The formal name for this algorithm, which was originally developed for
CSMA communication systems (the first was ALOHA, but it is also used by
e.g. Ethernet)[2] is "truncated binary exponential backoff"[3].

However, the term "exponential backoff" is already used to describe a
different algorithm in tenacity, that doesn't include jitter. These have
different use cases - geometrically increasing fixed delays are useful
for balancing retries against latency when a resource is unavailable and
you don't know when it will be available again; geometrically increasing
jitter is useful when multiple actors are in contention for the same
resource.

Since the "full_jitter" algorithm is effectively a convolution of the
"random" and "exponential" algorithms, "random_exponential" seems like a
better name for it.

The old name is still exported to avoid breaking any existing users.

[1] https://www.awsarchitectureblog.com/2015/03/backoff.html
[2] http://www.cs.utexas.edu/users/lam/NRL/backoff.html
[3] https://en.wikipedia.org/wiki/Exponential_backoff