Ensure sphinx can find tenacity when it runs autodoc
This means that building the documentation creates an API reference.

In the most recent build on ReadTheDocs [1], at the point where sphinx
actually runs, there are lots of errors from autodoc of the form:

    WARNING: autodoc: failed to import function 'retry' from module
        'tenacity'; the following exception was raised:
    No module named 'tenacity'

That's because Sphinx doesn't know how to find tenacity.  By adding the
directory containing tenacity to `sys.path`, autodoc is able to find
tenacity and create the API reference.

This is based on the template created by the sphinx-quickstart tool. [2]

[1]: https://readthedocs.org/projects/tenacity/builds/9021631/
[2]: https://github.com/sphinx-doc/sphinx/blob/a503ed8b781cd19f601b5294518aeb8adb9f011e/sphinx/templates/quickstart/conf.py_t#L9-L11