Add stack_info optional keyword parameter to before_sleep_log function to pass into logger
It may be desirable to see the current stack trace in the event of a retry, especially in the case of an exception.  exc_info only returns a trace up to the line that generates the exception, which may not be enough detail to determine the code path causing the retry.
- Updated test_before_sleep_log_raises_with_stack_info to use a more exact regular expression to capture the logger output.  Adding \Z at the end of the regex ensures the output logger text exactly matches what is expected since we're dealing with a multiline regex where '$' has a different meaning.
- Add additional test cases to ensure proper interoperability with exc_info